backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
python-dotenv==1.0.0
